move catalog to index
