+++ src/pages/index.tsx
@@ -1,120 +1,45 @@
-import styles from '@/styles/Home.module.css';
-import { Inter } from 'next/font/google';
+import { Anchor, Button, Card, Flex, Text, Title } from '@mantine/core';
 import Head from 'next/head';
-import Image from 'next/image';
+import Link from 'next/link';
+import { fetchItems } from './api';
+import { GuitarItem } from './types';
 
-const inter = Inter({ subsets: ['latin'] });
+type CatalogPageProps = { items: GuitarItem[] };
 
-export default function Home() {
+function CatalogPage({ items }: CatalogPageProps) {
   return (
     <>
       <Head>
-        <title>Create Next App</title>
+        <title>Catalog</title>
       </Head>
-      <main className={styles.main}>
-        <div className={styles.description}>
-          <p>
-            Get started by editing&nbsp;
-            <code className={styles.code}>src/pages/index.tsx</code>
-          </p>
-          <div>
-            <a
-              href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              By{' '}
-              <Image
-                src="/vercel.svg"
-                alt="Vercel Logo"
-                className={styles.vercelLogo}
-                width={100}
-                height={24}
-                priority
-              />
-            </a>
-          </div>
-        </div>
-
-        <div className={styles.center}>
-          <Image
-            className={styles.logo}
-            src="/next.svg"
-            alt="Next.js Logo"
-            width={180}
-            height={37}
-            priority
-          />
-          <div className={styles.thirteen}>
-            <Image
-              src="/thirteen.svg"
-              alt="13"
-              width={40}
-              height={31}
-              priority
-            />
-          </div>
-        </div>
-
-        <div className={styles.grid}>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Docs <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Find in-depth information about Next.js features and&nbsp;API.
-            </p>
-          </a>
-
-          <a
-            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Learn <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Learn about Next.js in an interactive course with&nbsp;quizzes!
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Templates <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Discover and deploy boilerplate example Next.js&nbsp;projects.
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Deploy <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Instantly deploy your Next.js site to a shareable URL
-              with&nbsp;Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
+      <Title>Catalog page</Title>
+      {items && (
+        <Flex wrap="wrap" gap="md">
+          {items.map(renderGuitarItem)}
+        </Flex>
+      )}
     </>
   );
+}
+
+function renderGuitarItem(item: GuitarItem) {
+  return (
+    <Card shadow="sm" p="lg" radius="md" withBorder key={item.uuid}>
+      <Anchor component={Link} href={{ pathname: item.uuid }}>
+        <Text size="lg">{item.name}</Text>
+      </Anchor>
+      <Button fullWidth mt="md">
+        Add to cart
+      </Button>
+    </Card>
+  );
+}
+
+export default CatalogPage;
+
+export async function getServerSideProps() {
+  const items = await fetchItems();
+  return {
+    props: { items }
+  };
 }
